[test][unit] Updated constant propagation and dfe tests
